// src/app/core/models/abastecimento.model.ts
export interface Abastecimento {
  id: number;
  data: string;
  posto: string;
  uf: string;
  tipoCombustivel: string;
  valorLitro: number;
  motorista: {
    nome: string;
    cpf: string;
  };
  veiculo: {
    placa: string;
    modelo: string;
  };
  quantidadeLitros: number;
  valorTotalPago: number;
}